fix: add tokenised font family support in input component
