+++ src/components/primitives/Input/InputBase.tsx
@@ -24,6 +24,7 @@ import { useToken } from '../../../hooks';
 import { usePropsResolution } from '../../../hooks/useThemeProps';
 import { useHover } from '@react-native-aria/interactions';
 import { mergeRefs } from '../../../utils';
+import { useResolvedFontFamily } from '../../../hooks/useResolvedFontFamily';
 
 const StyledInput = styled(TextInput)<IInputProps>(
   flex,
@@ -85,6 +86,9 @@ const InputBase = (
     _focus,
     _disabled,
     _invalid,
+    fontFamily,
+    fontWeight,
+    fontStyle,
     ...themedProps
   } = usePropsResolution('Input', {
     ...inputThemeProps,
@@ -94,9 +98,18 @@ const InputBase = (
   const _ref = React.useRef(null);
   const { isHovered } = useHover({}, _ref);
 
+  const resolvedFontFamily = useResolvedFontFamily({
+    fontFamily,
+    fontWeight,
+    fontStyle,
+  });
+
   return (
     <StyledInput
       {...inputProps}
+      fontFamily={resolvedFontFamily}
+      fontWeight={fontWeight}
+      fontStyle={fontStyle}
       secureTextEntry={type === 'password'}
       accessible
       accessibilityLabel={ariaLabel || accessibilityLabel}
